requirements.txt
# AI-First Smart Itinerary Planner - Python Dependencies

# Core Framework
langgraph>=0.0.50
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5

# LLM Providers
openai>=1.10.0

# Data Processing
pydantic>=2.0.0
python-dateutil>=2.8.2

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Logging & Monitoring
structlog>=23.1.0

# HTTP & APIs (for future integrations)
httpx>=0.25.0
requests>=2.31.0

# Database (for production)
# supabase>=2.0.0  # Uncomment for production
# psycopg2-binary>=2.9.0  # For PostgreSQL

# Utilities
typing-extensions>=4.8.0

# Development Dependencies
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.7.0
# flake8>=6.1.0
# mypy>=1.5.0

# Optional: For advanced features
# redis>=5.0.0  # For caching
# celery>=5.3.0  # For background tasks